fix(ReactViewManager): Add new base class for LinearGradientManager
ReactViewManager -> ReactViewManager + BorderedViewParentManager
LinearGradientPackage provides functionality to create views with linear gradients.

Fixes #284
